+++ src/lib/mrvEdit/mrvEditCallbacks.cpp
@@ -2887,9 +2887,18 @@ namespace mrv
             return;
         }
 
+        // Handle dragging the clip to the timeline viewport or over another
+        // clip.
         if (!file::isTemporaryEDL(destItem->path))
         {
-            LOG_ERROR(_("You can only add clips to an .otio EDL playlist."));
+            // 1. Get the index of the currently displayed clip (the 'A' item)
+            auto aIndex = ui->app->filesModel()->observeAIndex()->get();
+
+            // 2. Create a new timeline containing the currently displayed clip
+            add_clip_to_new_timeline_cb(aIndex, ui);
+
+            // 3. Append the dragged clip to this newly created timeline
+            add_clip_to_timeline_cb(index, ui);
             return;
         }